Move all the management of heat details to a new component called DetailsByHeat.
