+++ frontend/src/Event/EventDetails.jsx
@@ -2,21 +2,15 @@ import {
   ContentPaste as BackIcon,
   Build as BuildIcon,
   Download as DownloadIcon,
-  Edit as EditIcon,
-  Close as CloseIcon,
-  Save as SaveIcon,
 } from "@mui/icons-material";
 import { CircularProgress, Stack } from "@mui/material";
 import React, { useEffect, useState } from "react";
 import { SmmApi } from "../SmmApi.jsx";
 import AlertBox from "../components/Common/AlertBox.jsx";
-import ExpandableTable from "../components/Common/ExpandableTable";
 import ItemPaginationBar from "../components/Common/ItemPaginationBar";
 import TabPanel from "../components/Common/TabPanel";
-import HeatTimeForm from "../Heat/HeatTimeForm.jsx";
 import DetailsByLane from "../Heat/DetailsByLane.jsx";
-import { formatSeedTime } from "../utils/helperFunctions.js";
-import { useForm } from "react-hook-form";
+import DetailsByHeat from "../Heat/DetailsByHeat.jsx";
 
 const EventDetails = ({
   eventName,
@@ -36,10 +30,6 @@ const EventDetails = ({
   const [loading, setLoading] = useState(false);
   const [errorOnLoading, setErrorOnLoading] = useState(false);
   const [selectedTab, setSelectedTab] = useState(0);
-/*   const [editMainTableRowIds, setEditMainTableRowIds] = useState([]);
-  const laneFormHooks = Array.from({ length: numLanes }).map(() =>
-    useForm({ mode: "onChange" })
-  ); */
 
   let typeAlertLoading = errorOnLoading ? "error" : "success";
   let messageOnLoading = errorOnLoading
@@ -90,163 +80,15 @@ const EventDetails = ({
     },
   ];
 
-  //Need data for heat/lane tables
-  const mainHeatTableColumns = [
-    {
-      accessorKey: "heat_name",
-      header: "",
-      size: 200,
-      Cell: ({ cell }) => <strong>{cell.getValue()}</strong>,
-    },
-  ];
-
-/*   const mainLaneTableColumns = [
-    {
-      accessorKey: "lane_name",
-      header: "",
-      size: 200,
-      Cell: ({ cell }) => <strong>{cell.getValue()}</strong>,
-    },
-  ]; */
-
-  const getSubHeatTableColumns = (rowId) => [
-    {
-      accessorKey: "lane_num",
-      header: "Lane",
-      size: 50,
-    },
-    {
-      accessorKey: "athlete_full_name",
-      header: "Athlete",
-      size: 150,
-    },
-    {
-      accessorKey: "seed_time",
-      header: "Seed Time",
-      size: 100,
-      Cell: ({ cell }) => formatSeedTime(cell.getValue()),
-    },
-    {
-      accessorKey: "heat_time",
-      header: "Heat Time",
-      size: 100,
-      Cell: ({ cell }) => formatSeedTime(cell.getValue()),
-    },
-  ];
-
-/*   const getSubLaneTableColumns = (rowId) => {
-    const baseColumns = [
-      { accessorKey: "num_heat", header: "Heat", size: 50 },
-      { accessorKey: "athlete_full_name", header: "Athlete", size: 150 },
-    ];
-    const heatTimeColumn = {
-      accessorKey: "heat_time",
-      header: "Heat Time",
-      size: 100,
-      Cell: ({ cell }) => formatSeedTime(cell.getValue()),
-    };
-    const editHeatTimeColumn = {
-      accessorKey: "heat_time",
-      header: "Heat Time",
-      id: "edit_heat_time",
-      size: 250,
-      Cell: ({ row }) =>
-        row.original.athlete_full_name ? (
-          <HeatTimeForm
-            key={row.original.id}
-            control={laneFormHooks[rowId - 1]?.control}
-            name={String(row.original.id)}
-          />
-        ) : (
-          <br />
-        ),
-    };
-    return editMainTableRowIds.includes(rowId)
-      ? [baseColumns[0], baseColumns[1], editHeatTimeColumn]
-      : [baseColumns[0], baseColumns[1], heatTimeColumn];
-  }; */
-
-/*   const alreadyLaneUpdated = (rowIndex) => {
-    return laneData[rowIndex].heats.some((heat) => heat.heat_time !== null);
-  };
-
-  const handleEditClickOnMainTable = (rowIndex) => {
-    console.log("Row index:", rowIndex);
-    console.log("Data on that row:", laneData[rowIndex]);
-    setEditMainTableRowIds((prevIds) => {
-      const newIds = [...prevIds, laneData[rowIndex].id];
-      return newIds;
-    });
-  };
-
-  const handleSave = (rowIndex) => {
-    console.log("Save pressed.");
-    const laneFormState = laneFormHooks[rowIndex];
-    if (laneFormState) {
-      laneFormState.handleSubmit((data) => {
-        // TODO (Issue #150): Implement API call to persist the data
-        // Currently, 'data' is an object containing key-value pairs
-        // where each key represents a heat ID and the corresponding value
-        // is the entered heat time for that specific heat.
-        console.log("Submitting data for row:", rowIndex, "with values:", data);
-        laneFormState.reset();
-        setEditMainTableRowIds((prevIds) =>
-          prevIds.filter((id) => id !== laneData[rowIndex].id)
-        );
-      })();
-    } else {
-      console.log("No useForm");
-    }
-  };
-
-  const handleCloseClick = (rowIndex) => {
-    setEditMainTableRowIds((prevIds) =>
-      prevIds.filter((id) => id !== laneData[rowIndex].id)
-    );
-  };
-
-  const actionsLaneMainTable = [
-    {
-      name: "Edit",
-      icon: <EditIcon />,
-      onClick: handleEditClickOnMainTable,
-      tip: "Add/Edit heat time",
-      visible: (row) => {
-        const rowId = row.original.id;
-        const rowIndex = row.index;
-        return !(
-          editMainTableRowIds.includes(rowId) || alreadyLaneUpdated(rowIndex)
-        );
-      },
-    },
-    {
-      name: "Save",
-      icon: <SaveIcon />,
-      onClick: handleSave,
-      tip: "Save results",
-      visible: (row) => editMainTableRowIds.includes(row.original.id),
-    },
-    {
-      name: "Close",
-      icon: <CloseIcon />,
-      onClick: handleCloseClick,
-      tip: "Close edit mode",
-      visible: (row) => editMainTableRowIds.includes(row.original.id),
-    },
-  ]; */
-
   //Need for tabs component
   const tabs = [
     {
       label: "By Heat",
       content: (
-        <ExpandableTable
-          key={"heats" + eventId}
-          data={heatData}
-          columns={mainHeatTableColumns}
-          getSubTableColumns={getSubHeatTableColumns}
-          subData={"lanes"}
-        />
+        <DetailsByHeat
+        key={"heats" + eventId}
+        heatData={heatData}
+      />
       ),
     },
     {
